Extract the code that selects and fetchs App Experience data into reusable hooks.
